feat(TDI-41956) : fix TU & increase default timeout
